1.选项里面的填空

+++ app/src/main/java/com/doctor/sun/entity/Prescription.java
@@ -375,7 +375,7 @@ public class Prescription extends BaseObservable implements Parcelable, LayoutId
     }
 
     @Override
-    public float getSpan() {
+    public int getSpan() {
         return 12;
     }
