+++ app/src/main/java/com/doctor/sun/entity/Questions2.java
@@ -72,7 +72,7 @@ public class Questions2 implements SortedItem {
     }
 
     @Override
-    public float getSpan() {
+    public int getSpan() {
         return 12;
     }
